refactor-leave-structure combine a bunch of modules
