+++ src/init/initializeQuiz.js
@@ -1,33 +1,34 @@
 'use strict';
 
-import { QUESTION_CONTAINER_ID, QUIZ_CONTAINER_ID } from "../constants.js";
-import showCurrentQuestion from "../handlers/showCurrentQuestion.js";
-import createDOMElement from "../utils/createDOMElement.js";
-import getDOMElement from "../utils/getDOMElement.js";
-import createNextQuestionButtonElement from "../views/createNextQuestionButtonElement.js";
+import { QUESTION_CONTAINER_ID, QUIZ_CONTAINER_ID } from '../constants.js';
+import { showCurrentQuestion } from '../handlers/questionHandlers.js';
+import { createDOMElement, getDOMElement } from '../utils/DOMUtils.js';
+import { createNextQuestionButtonElement } from '../views/questionViews.js';
 import { quizData } from '../data.js';
 
 const initializeQuiz = () => {
-    quizData.currentQuestionIndex = 0;
+  quizData.currentQuestionIndex = 0;
 
-    console.log('bla');
-    setupQuizHTML();
+  console.log('bla');
+  setupQuizHTML();
 
-    showCurrentQuestion();
-}
+  showCurrentQuestion();
+};
 
 const setupQuizHTML = () => {
-    const userInterfaceContainer = getDOMElement('user-interface');
-    const quizContainer = createDOMElement('div', { id: QUIZ_CONTAINER_ID });
-    const questionContainer = createDOMElement('div', { id: QUESTION_CONTAINER_ID });
+  const userInterfaceContainer = getDOMElement('user-interface');
+  const quizContainer = createDOMElement('div', { id: QUIZ_CONTAINER_ID });
+  const questionContainer = createDOMElement('div', {
+    id: QUESTION_CONTAINER_ID,
+  });
 
-    quizContainer.appendChild(questionContainer);
+  quizContainer.appendChild(questionContainer);
 
-    const nextQuestionButton = createNextQuestionButtonElement();
-    quizContainer.appendChild(nextQuestionButton);
+  const nextQuestionButton = createNextQuestionButtonElement();
+  quizContainer.appendChild(nextQuestionButton);
 
-    console.log(quizContainer);
-    userInterfaceContainer.appendChild(quizContainer);
-}
+  console.log(quizContainer);
+  userInterfaceContainer.appendChild(quizContainer);
+};
 
 window.addEventListener('load', initializeQuiz);
