--- src/listeners/nextQuestion.js
@@ -1,9 +1,0 @@
-'use strict';
-
-import handleNextQuestion from "../handlers/handleNextQuestion.js";
-
-const nextQuestion = () => {
-    handleNextQuestion();
-}
-
-export default nextQuestion;
