+++ src/listeners/questionListeners.js
@@ -1,0 +1,7 @@
+'use strict';
+
+import { handleNextQuestion } from '../handlers/questionHandlers.js';
+
+export const nextQuestion = () => {
+  handleNextQuestion();
+};
