--- src/utils/clearDOMElement.js
@@ -1,10 +1,0 @@
-'use strict';
-
-/**
- * This function removes all the html inside the given element
- */
-const clearDOMElement = (DOMElement) => {
-    DOMElement.innerHTML = '';
-}
-
-export default clearDOMElement;
